DocumentPage: allow `document` being null.
In this case the page is not associated with a specific document. Instead any
empty document consisting only of the "AddPage" is shown. The only situation
where this can happen is if a new document is about to be created (but does not
exist, yet).

FossilOrigin-Name: b9802ebf5bd99e378535ae1fba83df4578aa3502