Implement thumbnails using urls to external files.
Instead of using the Thumbnail component we use the standard QML Image component
to showing the external thumbnail file. This is easier to integrate with QML and
allows not to expose the `Page` type QML. Furthermore, the interaction with the
`Document` as a list model is now more robust because only the url must be
returned instead of a pointer to the `Page`.

FossilOrigin-Name: 14a39bdb868dc829374ae0746fe0f510e256814b